Fix ITF14 checksum implementation
The sum of digits * their position value (1 or 3) sometimes lands on a multiple of 10 (e.g. 110).  The original implementation (`return 10 - (result % 10)`) would yield a checksum digit of `10`, where it should yield a `0`. Rounding up to the nearest 10 and subtracting by the result is the correct formula, as per [GS1's Specifications](http://www.gs1.org/how-calculate-check-digit-manually).